Conflicts:
	src/main/java/com/meta/overwash/mapper/LaundryMapper.java
	src/main/java/com/meta/overwash/mapper/ReservationMapper.java
	src/main/java/com/meta/overwash/service/LaundryService.java
	src/main/java/com/meta/overwash/service/LaundryServiceImpl.java
	src/main/resources/com/meta/overwash/mapper/LaundryMapper.xml
	src/main/resources/com/meta/overwash/mapper/ReservationMapper.xml
	src/main/webapp/WEB-INF/spring/root-context.xml
	src/main/webapp/WEB-INF/views/home.jsp
	src/main/webapp/WEB-INF/views/sign.jsp

+++ src/main/java/com/meta/overwash/domain/ReservationConfirmedDTO.java
@@ -12,7 +12,7 @@ import lombok.ToString;
 @NoArgsConstructor
 @ToString
 public class ReservationConfirmedDTO {
-	private Integer confirmId;
+	private Long confirmId;
 	private Date confirmDate;
 	private LaundryFactoryDTO factory;
 	private CrewDTO crew;
